cleanup types from mobile
