+++ packages/@coorpacademy-components/src/atom/choice/test/fixtures/disabled.tsx
@@ -1,5 +1,5 @@
 import noop from 'lodash/fp/noop';
-import {Media} from '../../../../types/progression-engine';
+import type {Media} from '../../../../molecule/questions/types';
 import {Props} from '../../index.native';
 
 type Fixture = {props: Props};
