+++ packages/@coorpacademy-components/src/molecule/questions/mobile/draggable/test/fixtures/choices-without-value.ts
@@ -1,6 +1,6 @@
 import noop from 'lodash/fp/noop';
 import {Props} from '../../index.native';
-import {Choice} from '../../../../../../types/progression-engine';
+import {Choice} from '../../../../types';
 
 type Fixture = {props: Props};
