--- packages/@coorpacademy-components/src/types/app-review.d.ts
@@ -1,3 +1,0 @@
-export type HandleBlur = () => void;
-export type HandleFocus = (id: string) => () => void;
-export type FocusedSelectId = string | null;
