+++ packages/@coorpacademy-components/src/types/styles.ts
